backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.100.0
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
google-generativeai==0.3.0